Load all projects as read-only in MSBuild.exe (#3584)
Uses XmlReaderSettings to ignore comments and whitespace for all projects. At the moment we read comments but then report the content as just String.Empty. The XmlReader still allocates a string for us which we immediately throw away.

This change adds a setting to ProjectCollection so that all projects loaded in the collection are treated as read-only. For command-line builds, a ProjectCollection is created for the build episode. End-users can create their own ProjectCollection with the setting as well.

Fixes #2576